packages/telegram-bot/requirements.txt
python-telegram-bot==20.3
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
